+++ source/psu/PSUController.hpp
@@ -23,27 +23,27 @@ public:
     void connectToPSU(msu_smdt::Port port);
     void disconnectFromPSU();
 
-    void powerOnChannels(std::vector<unsigned short> channels);
-    void powerOffChannels(std::vector<unsigned short> channels);
+    void powerOnChannels(std::vector<int> channels);
+    void powerOffChannels(std::vector<int> channels);
 
-    std::vector<float> getTestVoltages(std::vector<unsigned short> channels);
+    std::vector<float> getTestVoltages(std::vector<int> channels);
 
-    void setTestVoltages(std::vector<unsigned short> channels, float voltage);
-    void setTestCurrents(std::vector<unsigned short> channels, float current);
+    void setTestVoltages(std::vector<int> channels, float voltage);
+    void setTestCurrents(std::vector<int> channels, float current);
 
-    void setMaxVoltages(std::vector<unsigned short> channels, float voltage);
-    void setOverCurrentLimits(std::vector<unsigned short> channels, float time);
+    void setMaxVoltages(std::vector<int> channels, float voltage);
+    void setOverCurrentLimits(std::vector<int> channels, float time);
 
-    void setRampUpRate(std::vector<unsigned short> channels, float rate);
-    void setRampDownRate(std::vector<unsigned short> channels, float rate);
+    void setRampUpRate(std::vector<int> channels, float rate);
+    void setRampDownRate(std::vector<int> channels, float rate);
 
-    void killChannelsAfterTest(std::vector<unsigned short> channels, bool kill);
+    void killChannelsAfterTest(std::vector<int> channels, bool kill);
 
-    std::vector<float> readVoltages(std::vector<unsigned short> channels);
-    std::vector<float> readCurrents(std::vector<unsigned short> channels);
-    std::vector<unsigned long> readPolarities(std::vector<unsigned short> channels);
+    std::vector<float> readVoltages(std::vector<int> channels);
+    std::vector<float> readCurrents(std::vector<int> channels);
+    std::vector<unsigned long> readPolarities(std::vector<int> channels);
 
-    std::vector<unsigned long> readStatuses(std::vector<unsigned short> channels);
+    std::vector<unsigned long> readStatuses(std::vector<int> channels);
 
 private:
     bool forceClosed;
